Disable PRs for minor/patch updates
They are too noisy and normally not necessary:

- For minor updates, it's normally not necessary to need any new
  features.

- For patch updates, the assumption is most of the time the
  bug/security fixes are minor or not affecting the project.

Note lock files updates already update any minor/patch version if the
appropriate dependency version ranges are used (eg, ^1.2.3), which is
the typical case.

Note the 0.x case (major versions ressembling minor versions) is not handled as
it shouldn't be that common nor necessary.